Create components for each section in the app, i've created custom hooks for handle data from API routes
